About dialog: various improvements.
