+++ src/renderer/src/App.vue
@@ -1,11 +1,20 @@
 <template>
   <Background />
   <ResetAllDialog />
-  <AboutDialog :version="version" :copyright="copyright" />
+  <AboutDialog v-model:visible="visible" @close="visible = false" />
 </template>
 
 <script setup lang="ts">
-const version = __APP_VERSION__
-const currentYear = new Date().getFullYear()
-const copyright = currentYear === 2024 ? '2024' : `2024-${currentYear}`
+import { ref } from 'vue'
+
+const visible = ref(false)
+
+// @ts-ignore (window.electronAPI may or not be defined and that is why we test it)
+const electronAPI = window.electronAPI
+
+if (electronAPI !== undefined) {
+  electronAPI.onAbout(() => {
+    visible.value = true
+  })
+}
 </script>
